create custom err handler

+++ src/app/controllers/deleteUser.js
@@ -1,34 +1,39 @@
 var { dbServices } = require("../../database/db");
-var { reportError } = require("../../bot/errReportBot");
+var { DeleteUserError } = require("../customError/index");
 var validateAuthHeader = require("../services/validateAuthHeader");
 
-var deleteUser = async (req, res) => {
+var deleteUser = async (req, res, next) => {
   try {
     var authHeader = req.headers?.authorization;
 
+    if (!authHeader) {
+      return res.sendStatus(401);
+    }
     var validAuthHeader = await validateAuthHeader(authHeader);
 
     if (!validAuthHeader) {
-      return res.sendStatus(401);
+      return res.sendStatus(403);
     }
 
     var db = await dbServices();
 
     var { userId } = req.body;
 
+    var user = await db.getUserById(userId);
+
+    if (!user) {
+      return res.sendStatus(404);
+    }
+
     var isUserDeleted = await db.deleteUser(userId);
 
     if (!isUserDeleted) {
-      return res.sendStatus(304);
+      throw new DeleteUserError();
     }
 
     return res.sendStatus(200);
   } catch (err) {
-    err.location = "deleteUser controller";
-
-    await reportError(userId, err);
-
-    return res.sendStatus(500);
+    next(err);
   }
 };
